+++ clikt/src/commonMain/kotlin/com/github/ajalt/clikt/command/ChainedCliktCommand.kt
@@ -9,28 +9,12 @@ import com.github.ajalt.clikt.parsers.CommandLineParser
  * subcommands.
  */
 abstract class ChainedCliktCommand<T>(
-    help: String = "",
-    epilog: String = "",
+    /**
+     * The name of the program to use in the help output. If not given, it is inferred from the
+     * class name.
+     */
     name: String? = null,
-    invokeWithoutSubcommand: Boolean = false,
-    printHelpOnEmptyArgs: Boolean = false,
-    helpTags: Map<String, String> = emptyMap(),
-    autoCompleteEnvvar: String? = "",
-    allowMultipleSubcommands: Boolean = false,
-    treatUnknownOptionsAsArgs: Boolean = false,
-    hidden: Boolean = false,
-) : BaseCliktCommand<ChainedCliktCommand<T>>(
-    help,
-    epilog,
-    name,
-    invokeWithoutSubcommand,
-    printHelpOnEmptyArgs,
-    helpTags,
-    autoCompleteEnvvar,
-    allowMultipleSubcommands,
-    treatUnknownOptionsAsArgs,
-    hidden
-) {
+) : BaseCliktCommand<ChainedCliktCommand<T>>(name) {
     /**
      * Perform actions after parsing is complete and this command is invoked.
      *