+++ clikt/src/commonMain/kotlin/com/github/ajalt/clikt/completion/CompletionBuiltins.kt
@@ -2,6 +2,7 @@ package com.github.ajalt.clikt.completion
 
 import com.github.ajalt.clikt.core.BaseCliktCommand
 import com.github.ajalt.clikt.core.CliktCommand
+import com.github.ajalt.clikt.core.Context
 import com.github.ajalt.clikt.parameters.arguments.argument
 import com.github.ajalt.clikt.parameters.options.option
 import com.github.ajalt.clikt.parameters.options.validate
@@ -17,8 +18,10 @@ fun <T : BaseCliktCommand<*>> T.completionOption(
     help: String = "",
     hidden: Boolean = false,
 ): T = apply {
-    registerOption(option(*names, help = help, hidden = hidden, eager = true,
-        metavar = choices.joinToString("|", prefix = "(", postfix = ")")).validate {
+    registerOption(option(
+        *names, help = help, hidden = hidden, eager = true,
+        metavar = choices.joinToString("|", prefix = "(", postfix = ")")
+    ).validate {
         CompletionGenerator.throwCompletionMessage(context.command, it)
     })
 }
@@ -27,10 +30,12 @@ fun <T : BaseCliktCommand<*>> T.completionOption(
  * A subcommand that will print a completion script for the given shell when invoked.
  */
 class CompletionCommand(
-    help: String = "Generate a tab-complete script for the given shell",
-    epilog: String = "",
+    private val help: String = "Generate a tab-complete script for the given shell",
+    private val epilog: String = "",
     name: String = "generate-completion",
-) : CliktCommand(help, epilog, name) {
+) : CliktCommand(name) {
+    override fun commandHelp(context: Context): String = help
+    override fun commandHelpEpilog(context: Context): String = epilog
     private val shell by argument("shell").choice(*choices)
     override fun run() {
         val cmd = currentContext.parent?.command ?: this