+++ samples/copy/src/main/kotlin/com/github/ajalt/clikt/samples/copy/main.kt
@@ -1,6 +1,7 @@
 package com.github.ajalt.clikt.samples.copy
 
 import com.github.ajalt.clikt.core.CliktCommand
+import com.github.ajalt.clikt.core.Context
 import com.github.ajalt.clikt.core.main
 import com.github.ajalt.clikt.core.terminal
 import com.github.ajalt.clikt.parameters.arguments.argument
@@ -10,7 +11,11 @@ import com.github.ajalt.clikt.parameters.options.option
 import com.github.ajalt.clikt.parameters.types.file
 import com.github.ajalt.mordant.terminal.YesNoPrompt
 
-class Copy : CliktCommand(help = "Copy SOURCE to DEST, or multiple SOURCE(s) to directory DEST.") {
+class Copy : CliktCommand() {
+    override fun commandHelp(context: Context): String {
+        return "Copy SOURCE to DEST, or multiple SOURCE(s) to directory DEST."
+    }
+
     val interactive by option("-i", "--interactive", help = "prompt before overwrite").flag()
     val recursive by option("-r", "--recursive", help = "copy directories recursively").flag()
     val source by argument().file(mustExist = true).multiple()