+++ samples/plugins/src/main/kotlin/com/github/ajalt/clikt/samples/plugins/delete.kt
@@ -1,18 +1,20 @@
 package com.github.ajalt.clikt.samples.plugins
 
 import com.github.ajalt.clikt.core.CliktCommand
+import com.github.ajalt.clikt.core.Context
 import com.github.ajalt.clikt.core.requireObject
 import org.kodein.di.Kodein
 import org.kodein.di.generic.bind
 import org.kodein.di.generic.inSet
 import org.kodein.di.generic.provider
 
-class Delete : CliktCommand(
-    help = """
-        Deletes a repository.
+class Delete : CliktCommand() {
+    override fun commandHelp(context: Context) = """
+    Deletes a repository.
 
-        This will throw away the current repository.
-        """.trimIndent()) {
+    This will throw away the current repository.
+    """.trimIndent()
+
     val repo: Repo by requireObject()
 
     override fun run() {